Move BuildStatus to constants module
