+++ build-coordinator/src/main/java/org/jboss/pnc/coordinator/builder/BuildTasksInitializer.java
@@ -223,7 +223,7 @@ public class BuildTasksInitializer {
                 .buildConfigurationSet(buildConfigurationSet)
                 .user(user)
                 .startTime(new Date())
-                .status(org.jboss.pnc.model.BuildStatus.BUILDING)
+                .status(org.jboss.pnc.enums.BuildStatus.BUILDING)
                 .temporaryBuild(buildOptions.isTemporaryBuild())
                 .build();
