+++ build-coordinator/src/main/java/org/jboss/pnc/coordinator/builder/datastore/DatastoreAdapter.java
@@ -24,7 +24,7 @@ import org.jboss.pnc.model.BuildConfiguration;
 import org.jboss.pnc.model.BuildConfigurationAudited;
 import org.jboss.pnc.model.BuildConfigurationSet;
 import org.jboss.pnc.model.BuildRecord;
-import org.jboss.pnc.model.BuildStatus;
+import org.jboss.pnc.enums.BuildStatus;
 import org.jboss.pnc.model.Project;
 import org.jboss.pnc.spi.BuildOptions;
 import org.jboss.pnc.spi.BuildResult;
@@ -42,6 +42,7 @@ import org.slf4j.LoggerFactory;
 
 import javax.enterprise.context.Dependent;
 import javax.inject.Inject;
+
 import java.io.PrintWriter;
 import java.io.StringWriter;
 import java.time.Instant;
@@ -53,11 +54,11 @@ import java.util.Optional;
 import java.util.Set;
 import java.util.stream.Collectors;
 
-import static org.jboss.pnc.model.BuildStatus.CANCELLED;
-import static org.jboss.pnc.model.BuildStatus.FAILED;
-import static org.jboss.pnc.model.BuildStatus.REJECTED;
-import static org.jboss.pnc.model.BuildStatus.SYSTEM_ERROR;
-import static org.jboss.pnc.model.BuildStatus.UNKNOWN;
+import static org.jboss.pnc.enums.BuildStatus.CANCELLED;
+import static org.jboss.pnc.enums.BuildStatus.FAILED;
+import static org.jboss.pnc.enums.BuildStatus.REJECTED;
+import static org.jboss.pnc.enums.BuildStatus.SYSTEM_ERROR;
+import static org.jboss.pnc.enums.BuildStatus.UNKNOWN;
 
 /**
  * Created by <a href="mailto:matejonnet@gmail.com">Matej Lazar</a> on 2014-12-15.
