+++ model/src/main/java/org/jboss/pnc/model/BuildConfigSetRecord.java
@@ -17,6 +17,7 @@
  */
 package org.jboss.pnc.model;
 
+import org.jboss.pnc.enums.BuildStatus;
 import org.hibernate.annotations.ForeignKey;
 import org.hibernate.annotations.Index;
 
@@ -38,6 +39,7 @@ import javax.persistence.PersistenceException;
 import javax.persistence.PreRemove;
 import javax.persistence.SequenceGenerator;
 import javax.validation.constraints.NotNull;
+
 import java.util.Date;
 import java.util.HashMap;
 import java.util.HashSet;
