+++ rest-model/src/main/java/org/jboss/pnc/rest/restmodel/bpm/BuildResultRest.java
@@ -21,9 +21,10 @@ package org.jboss.pnc.rest.restmodel.bpm;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
+
 import org.jboss.pnc.common.util.StringUtils;
 import org.jboss.pnc.model.Artifact;
-import org.jboss.pnc.model.BuildStatus;
+import org.jboss.pnc.enums.BuildStatus;
 import org.jboss.pnc.rest.restmodel.BuildDriverResultRest;
 import org.jboss.pnc.rest.restmodel.BuildExecutionConfigurationRest;
 import org.jboss.pnc.rest.restmodel.RepositoryManagerResultRest;
@@ -40,6 +41,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import javax.xml.bind.annotation.XmlRootElement;
+
 import java.io.IOException;
 import java.io.Serializable;
 import java.util.List;
