+++ spi/src/main/java/org/jboss/pnc/spi/coordinator/BuildSetTask.java
@@ -18,11 +18,11 @@
 package org.jboss.pnc.spi.coordinator;
 
 import lombok.Getter;
+
 import org.jboss.pnc.model.BuildConfigSetRecord;
-import org.jboss.pnc.model.BuildConfiguration;
 import org.jboss.pnc.model.BuildConfigurationAudited;
-import org.jboss.pnc.model.BuildStatus;
-import org.jboss.pnc.spi.BuildCoordinationStatus;
+import org.jboss.pnc.enums.BuildCoordinationStatus;
+import org.jboss.pnc.enums.BuildStatus;
 import org.jboss.pnc.spi.BuildOptions;
 import org.jboss.pnc.spi.BuildSetStatus;
 import org.slf4j.Logger;
